New Test Structure for Reactive Apps.
