+++ src/test/java/me/alidg/errors/reactive/ReactiveApplication.java
@@ -1,0 +1,12 @@
+package me.alidg.errors.reactive;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReactiveApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ReactiveApplication.class);
+    }
+}
