+++ src/test/java/me/alidg/errors/reactive/ReactiveIT.java
@@ -4,64 +4,28 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 import me.alidg.errors.ExceptionLogger;
-import me.alidg.errors.HttpError;
 import me.alidg.errors.HttpError.CodedMessage;
-import me.alidg.errors.WebErrorHandlerPostProcessor;
-import me.alidg.errors.annotation.ExceptionMapping;
-import me.alidg.errors.annotation.ExposeAsArg;
-import me.alidg.errors.conf.ErrorsAutoConfiguration;
-import me.alidg.errors.conf.ReactiveErrorsAutoConfiguration;
-import me.alidg.errors.conf.ReactiveSecurityErrorsAutoConfiguration;
 import me.alidg.errors.handlers.ServletWebErrorHandler;
 import org.hamcrest.Matchers;
-import org.junit.After;
-import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.ArgumentCaptor;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.WebApplicationType;
-import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
-import org.springframework.boot.autoconfigure.context.MessageSourceAutoConfiguration;
-import org.springframework.boot.autoconfigure.context.PropertyPlaceholderAutoConfiguration;
-import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
-import org.springframework.boot.autoconfigure.security.reactive.ReactiveSecurityAutoConfiguration;
-import org.springframework.boot.autoconfigure.validation.ValidationAutoConfiguration;
-import org.springframework.boot.autoconfigure.web.reactive.HttpHandlerAutoConfiguration;
-import org.springframework.boot.autoconfigure.web.reactive.ReactiveWebServerFactoryAutoConfiguration;
-import org.springframework.boot.autoconfigure.web.reactive.WebFluxAutoConfiguration;
-import org.springframework.boot.autoconfigure.web.reactive.error.ErrorWebFluxAutoConfiguration;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.test.util.ApplicationContextTestUtils;
-import org.springframework.boot.web.reactive.context.ConfigurableReactiveWebApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.MediaType;
 import org.springframework.http.client.MultipartBodyBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableReactiveMethodSecurity;
-import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
-import org.springframework.security.config.web.server.ServerHttpSecurity;
-import org.springframework.security.web.server.SecurityWebFilterChain;
-import org.springframework.security.web.server.ServerAuthenticationEntryPoint;
-import org.springframework.security.web.server.authorization.ServerAccessDeniedHandler;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit4.rules.SpringClassRule;
+import org.springframework.test.context.junit4.rules.SpringMethodRule;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.util.MultiValueMap;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.reactive.config.EnableWebFlux;
-import org.springframework.web.reactive.config.WebFluxConfigurationSupport;
-import org.springframework.web.reactive.function.server.ServerRequest;
-import reactor.core.publisher.Mono;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 import java.util.Locale;
 import java.util.stream.Stream;
 
@@ -72,56 +36,45 @@ import static me.alidg.errors.handlers.MissingRequestParametersWebErrorHandler.*
 import static me.alidg.errors.handlers.ServletWebErrorHandler.*;
 import static me.alidg.errors.handlers.SpringSecurityWebErrorHandler.ACCESS_DENIED;
 import static me.alidg.errors.handlers.SpringSecurityWebErrorHandler.AUTH_REQUIRED;
-import static me.alidg.errors.reactive.ReactiveIT.Dto.dto;
-import static org.assertj.core.api.Assertions.assertThat;
+import static me.alidg.errors.reactive.ReactiveController.Dto.dto;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.springframework.http.HttpHeaders.ACCEPT_LANGUAGE;
-import static org.springframework.http.HttpMethod.GET;
-import static org.springframework.http.HttpMethod.POST;
 import static org.springframework.http.HttpStatus.METHOD_NOT_ALLOWED;
 import static org.springframework.http.HttpStatus.NOT_ACCEPTABLE;
 import static org.springframework.http.HttpStatus.*;
 import static org.springframework.http.MediaType.*;
-import static org.springframework.security.test.web.reactive.server.SecurityMockServerConfigurers.*;
+import static org.springframework.security.test.web.reactive.server.SecurityMockServerConfigurers.csrf;
+import static org.springframework.security.test.web.reactive.server.SecurityMockServerConfigurers.mockUser;
 
 /**
  * Integration tests for our reactive stack support.
  *
  * @author Ali Dehghani
  */
+@AutoConfigureWebTestClient
 @RunWith(JUnitParamsRunner.class)
+@SpringBootTest(classes = ReactiveApplication.class)
+@TestPropertySource(properties = {
+    "errors.add-fingerprint=false",
+    "errors.expose-arguments=non_empty",
+    "spring.main.allow-bean-definition-overriding=true",
+    "spring.main.web-application-type=reactive"
+})
 public class ReactiveIT {
 
+    @ClassRule
+    public static final SpringClassRule springClassRule = new SpringClassRule();
+
+    @Rule
+    public final SpringMethodRule springMethodRule = new SpringMethodRule();
+
+    @Autowired
     private WebTestClient client;
-    private ConfigurableReactiveWebApplicationContext context;
 
-    @Before
-    public void setUp() {
-        System.setProperty("errors.expose-arguments", "non_empty");
-        System.setProperty("errors.add-fingerprint", "false");
-        String port = "--server.port=0";
-        String messages = "--spring.messages.basename=test_messages";
-        SpringApplication application = new SpringApplication(WebFluxConfig.class);
-        application.setWebApplicationType(WebApplicationType.REACTIVE);
-
-        context = (ConfigurableReactiveWebApplicationContext) application.run(port, messages);
-
-        client = WebTestClient
-            .bindToApplicationContext(context)
-            .apply(springSecurity())
-            .configureClient()
-            .build();
-    }
-
-    @After
-    public void tearDown() {
-        System.clearProperty("errors.expose-arguments");
-        System.clearProperty("errors.add-fingerprint");
-        ApplicationContextTestUtils.closeAll(context);
-    }
+    @MockBean
+    private ExceptionLogger logger;
 
     @Test
     public void errorAttributes_ShouldBeAbleToHandleUnknownErrors() {
@@ -132,7 +85,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments").doesNotExist();
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -147,7 +100,7 @@ public class ReactiveIT {
             .jsonPath("$.errors[0].arguments.c").isEqualTo("c")
             .jsonPath("$.errors[0].arguments.f").isEqualTo("10");
 
-        verify(logger()).log(any(ReactiveIT.InvalidParamsException.class));
+        verify(logger).log(any(ReactiveController.InvalidParamsException.class));
     }
 
     @Test
@@ -173,7 +126,7 @@ public class ReactiveIT {
             .jsonPath("errors[*].message").value(Matchers.containsInAnyOrder(messages))
             .jsonPath("$.fingerprint").doesNotExist();
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -185,7 +138,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments.path").isEqualTo("/should_not_be_found");
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -197,7 +150,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments").doesNotExist();
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -209,7 +162,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments").doesNotExist();
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -221,7 +174,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments").exists();
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -230,7 +183,7 @@ public class ReactiveIT {
             .expectStatus().isEqualTo(UNSUPPORTED_MEDIA_TYPE)
             .expectBody().jsonPath("errors[0].code").isEqualTo(NOT_SUPPORTED);
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -238,7 +191,7 @@ public class ReactiveIT {
         client.get().uri("/test/param?name=ali").accept(IMAGE_JPEG).exchange()
             .expectStatus().isEqualTo(NOT_ACCEPTABLE);
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -251,7 +204,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments.method").isEqualTo("PUT");
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -265,7 +218,7 @@ public class ReactiveIT {
             .jsonPath("$.errors[0].arguments.name").isEqualTo("name")
             .jsonPath("$.errors[0].arguments.expected").isEqualTo("String");
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -279,7 +232,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments.name").isEqualTo("file");
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -291,7 +244,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments").doesNotExist();
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -305,7 +258,7 @@ public class ReactiveIT {
             .jsonPath("$.fingerprint").doesNotExist()
             .jsonPath("$.errors[0].arguments").doesNotExist();
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -318,7 +271,7 @@ public class ReactiveIT {
             .jsonPath("$.errors[0].arguments.name").isEqualTo("name")
             .jsonPath("$.errors[0].arguments.expected").isEqualTo("String");
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -331,7 +284,7 @@ public class ReactiveIT {
             .jsonPath("$.errors[0].arguments.name").isEqualTo("name")
             .jsonPath("$.errors[0].arguments.expected").isEqualTo("String");
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -344,7 +297,7 @@ public class ReactiveIT {
             .jsonPath("$.errors[0].arguments.name").isEqualTo("name")
             .jsonPath("$.errors[0].arguments.expected").isEqualTo("String");
 
-        verify(logger()).log(any());
+        verify(logger).log(any());
     }
 
     @Test
@@ -373,32 +326,11 @@ public class ReactiveIT {
             .jsonPath("$.errors[0].arguments.expected").isEqualTo("Integer");
     }
 
-    @Test
-    public void errorAttributes_HttpErrorShouldContainAppropriateDetails() {
-        WebErrorHandlerPostProcessor processor = context.getBean(WebErrorHandlerPostProcessor.class);
-        ArgumentCaptor<HttpError> captor = ArgumentCaptor.forClass(HttpError.class);
-
-        client.get().uri("/test/type-mismatch?number=invalid")
-            .exchange()
-            .expectStatus().isBadRequest();
-
-        verify(processor).process(captor.capture());
-        HttpError value = captor.getValue();
-        assertThat(value.getRequest()).isNotNull();
-        assertThat(value.getRequest()).isInstanceOf(ServerRequest.class);
-        assertThat(value.getOriginalException()).isNotNull();
-        assertThat(value.getRefinedException()).isNotNull();
-    }
-
     private MultiValueMap<String, HttpEntity<?>> generateBody() {
         MultipartBodyBuilder builder = new MultipartBodyBuilder();
         builder.part("fieldPart", "fieldValue");
         builder.part("fileParts", new ByteArrayResource("data".getBytes()));
         return builder.build();
-    }
-
-    private ExceptionLogger logger() {
-        return context.getBean(ExceptionLogger.class);
     }
 
     private Object[] provideInvalidBody() {
@@ -421,220 +353,5 @@ public class ReactiveIT {
 
     private CodedMessage cm(String code, String message) {
         return new CodedMessage(code, message, emptyList());
-    }
-
-    protected enum Sort {
-        ASC, DESC
-    }
-
-    @RestController
-    @RequestMapping("/test")
-    protected static class ReactiveTestController {
-
-        @GetMapping
-        public Mono<Void> get() {
-            throw new ReactiveIT.InvalidParamsException(10, "c", "a");
-        }
-
-        @PostMapping
-        public Mono<Void> post(@RequestBody @Validated ReactiveIT.Dto dto) {
-            return Mono.empty();
-        }
-
-        @DeleteMapping
-        public Mono<Void> delete() {
-            throw new IllegalArgumentException();
-        }
-
-        @GetMapping("/param")
-        public Mono<Dto> getParam(@RequestParam String name) {
-            return Mono.just(new ReactiveIT.Dto(name, 12, ""));
-        }
-
-        @PostMapping(value = "/part", consumes = MULTIPART_FORM_DATA_VALUE)
-        public MultipartFile postParam(@RequestPart MultipartFile file) {
-            return file;
-        }
-
-        @GetMapping("/protected")
-        public Mono<Void> needsAuthentication() {
-            return Mono.empty();
-        }
-
-        @PostMapping("/protected")
-        public Mono<Void> needsPermission() {
-            return Mono.empty();
-        }
-
-        @GetMapping("/header")
-        public Mono<Void> headerIsRequired(@RequestHeader String name) {
-            return Mono.empty();
-        }
-
-        @GetMapping("/cookie")
-        public Mono<Void> cookieIsRequired(@CookieValue String name) {
-            return Mono.empty();
-        }
-
-        @GetMapping("/matrix")
-        public Mono<Void> matrixIsRequired(@MatrixVariable String name) {
-            return Mono.empty();
-        }
-
-        @GetMapping("/type-mismatch")
-        public void mismatch(@RequestParam Integer number) {
-        }
-
-        @GetMapping("/paged")
-        public void pagedResult(Pageable pageable) {
-        }
-    }
-
-    protected static class Dto {
-
-        @NotBlank(message = "{text.required}")
-        private String text;
-
-        @Min(value = 0, message = "number.min")
-        private int number;
-
-        @Size(min = 1, max = 2, message = "range.limit")
-        private List<String> range;
-
-        Dto() {
-        }
-
-        Dto(String text, int number, String... range) {
-            this.text = text;
-            this.number = number;
-            this.range = Arrays.asList(range);
-        }
-
-        static ReactiveIT.Dto dto(String text, int number, String... range) {
-            return new ReactiveIT.Dto(text, number, range);
-        }
-
-        public String getText() {
-            return text;
-        }
-
-        public void setText(String text) {
-            this.text = text;
-        }
-
-        public int getNumber() {
-            return number;
-        }
-
-        public void setNumber(int number) {
-            this.number = number;
-        }
-
-        public List<String> getRange() {
-            return range;
-        }
-
-        public void setRange(List<String> range) {
-            this.range = range;
-        }
-    }
-
-    protected static class Pageable {
-
-        private Integer page;
-        private Integer size;
-        private Sort sort;
-
-        public Integer getPage() {
-            return page;
-        }
-
-        public void setPage(Integer page) {
-            this.page = page;
-        }
-
-        public Integer getSize() {
-            return size;
-        }
-
-        public void setSize(Integer size) {
-            this.size = size;
-        }
-
-        public Sort getSort() {
-            return sort;
-        }
-
-        public void setSort(Sort sort) {
-            this.sort = sort;
-        }
-    }
-
-    @EnableWebFlux
-    @TestConfiguration
-    @EnableWebFluxSecurity
-    @ImportAutoConfiguration({
-        ErrorsAutoConfiguration.class,
-        JacksonAutoConfiguration.class,
-        WebFluxAutoConfiguration.class,
-        ValidationAutoConfiguration.class,
-        HttpHandlerAutoConfiguration.class,
-        ErrorWebFluxAutoConfiguration.class,
-        MessageSourceAutoConfiguration.class,
-        ReactiveErrorsAutoConfiguration.class,
-        ReactiveSecurityAutoConfiguration.class,
-        ReactiveSecurityErrorsAutoConfiguration.class,
-        PropertyPlaceholderAutoConfiguration.class,
-        ReactiveWebServerFactoryAutoConfiguration.class,
-    })
-    @EnableReactiveMethodSecurity
-    @ComponentScan(basePackageClasses = ReactiveIT.ReactiveTestController.class)
-    static class WebFluxConfig extends WebFluxConfigurationSupport {
-
-        @Bean
-        public SecurityWebFilterChain securityWebFilterChain(ServerHttpSecurity http,
-                                                             ServerAccessDeniedHandler accessDeniedHandler,
-                                                             ServerAuthenticationEntryPoint authenticationEntryPoint) {
-            return http
-                .csrf()
-                .accessDeniedHandler(accessDeniedHandler)
-                .and()
-                .exceptionHandling()
-                .authenticationEntryPoint(authenticationEntryPoint)
-                .accessDeniedHandler(accessDeniedHandler)
-                .and()
-                .authorizeExchange()
-                .pathMatchers(GET, "/test/protected").authenticated()
-                .pathMatchers(POST, "/test/protected").hasRole("ADMIN")
-                .anyExchange().permitAll()
-                .and().build();
-        }
-
-        @Bean
-        public ExceptionLogger exceptionLogger() {
-            return mock(ExceptionLogger.class);
-        }
-
-        @Bean
-        public WebErrorHandlerPostProcessor postProcessor() {
-            return mock(WebErrorHandlerPostProcessor.class);
-        }
-    }
-
-    @ExceptionMapping(statusCode = UNPROCESSABLE_ENTITY, errorCode = "invalid_params")
-    private static class InvalidParamsException extends RuntimeException {
-
-        @ExposeAsArg(10)
-        private final int f;
-        @ExposeAsArg(2)
-        private final String c;
-        @ExposeAsArg(0)
-        private final String a;
-
-        InvalidParamsException(int f, String c, String a) {
-            this.f = f;
-            this.c = c;
-            this.a = a;
-        }
     }
 }
